refactor: Moving NamedControls and NamedComponents into folders
