+++ QscQsys/QscQsys/NamedComponents/QsysRouter.cs
@@ -3,7 +3,7 @@ using Crestron.SimplSharp;
 using QscQsys.Intermediaries;
 using QscQsys.Utils;
 
-namespace QscQsys
+namespace QscQsys.NamedComponents
 {
     public sealed class QsysRouter : QsysComponent
     {
