+++ QscQsys/QscQsys/NamedComponents/QsysSoftphoneController.cs
@@ -1,4 +1,4 @@
-﻿namespace QscQsys
+﻿namespace QscQsys.NamedComponents
 {
     public class QsysSoftphoneController : QsysPotsController
     {
